refactor: changed some names based on naming conventions
