+++ src/components/metrics-charts/StackedBarCharMetric.tsx
