+++ src/components/pages-header/index.tsx
@@ -1,0 +1,3 @@
+import { PagesHeader } from '@/components/pages-header/PagesHeader';
+
+export { PagesHeader };
